+++ core/abdeveng_core/abdeveng_core_as7.componentinfo.xml
