+++ core/abdeveng_core/core_vs2013.vcxproj.filters
