+++ core/abdeveng_core/include/abdeveng/core/logging/common/logger.h
