+++ core/abdeveng_core/src/abdeveng/core/logging/common/logging.cpp
