+++ core/abdeveng_core_as7.atsln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Atmel Studio Solution File, Format Version 11.00
 VisualStudioVersion = 14.0.23107.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{E66E83B9-2572-4076-B26E-6BE79FF3018A}") = "core_as7", "core\core_as7.cppproj", "{24B3129D-8A38-4C38-8819-D6F354762A5D}"
+Project("{E66E83B9-2572-4076-B26E-6BE79FF3018A}") = "abdeveng_core_as7", "abdeveng_core\abdeveng_core_as7.cppproj", "{24B3129D-8A38-4C38-8819-D6F354762A5D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
