+++ core/abdeveng_core_vs2013.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
 VisualStudioVersion = 12.0.40629.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "core_vs2013", "core\core_vs2013.vcxproj", "{D8683758-AAC9-4C64-A180-75FB7212010E}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "abdeveng_core_vs2013", "abdeveng_core\core_vs2013.vcxproj", "{D8683758-AAC9-4C64-A180-75FB7212010E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
